Split out logic for storing data into separate mdule
